Added instantiate event method

+++ TuPencaUy.CoreAPI/TuPencaUy.CoreAPI/Controllers/EventController.cs
@@ -1,5 +1,6 @@
 ﻿using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
+using System.ComponentModel.DataAnnotations;
 using System.Net;
 using TuPencaUy.Core.API.Model.Requests;
 using TuPencaUy.Core.API.Model.Responses;
@@ -16,10 +17,54 @@ namespace TuPencaUy.Core.API.Controllers
   public class EventController : BaseController
   {
     private readonly IEventService _eventService;
+    private readonly IServiceFactory _serviceFactory;
 
     public EventController(IServiceFactory serviceFactory)
     {
       _eventService = serviceFactory.GetService<IEventService>();
+      _serviceFactory = serviceFactory;
+    }
+
+    [HttpPost("InstantiateEvent")]
+    public IActionResult InstantiateEvent([Required] int eventId)
+    {
+      try
+      {
+        string domain = ObtainTenantFromToken();
+        if (string.IsNullOrEmpty(domain))
+        {
+          return BadRequest(new ApiResponse { Error = true, Message = "You must be logged to a tenant" });
+        }
+
+        _serviceFactory.CreatePlatformServices();
+        var platformEventService = _serviceFactory.GetService<IEventService>();
+
+        var searchedEvent = platformEventService.GetEvent(eventId);
+        var searchedMatches = platformEventService.GetMatches(
+          out int count,
+          idTeam: null,
+          otherIdTeam: null,
+          eventId: eventId,
+          sportId: null,
+          fromDate: null,
+          untilDate: null,
+          page: null,
+          pageSize: int.MaxValue);
+
+        string connStringTennant = _serviceFactory.GetService<ISiteService>().GetSiteByDomain(domain).ConnectionString;
+
+        _serviceFactory.CreateTenantServices(connStringTennant);
+
+        var newEvent = _serviceFactory.GetService<IEventService>().InstantiateEvent(searchedEvent, searchedMatches);
+
+        var data = new { Event = newEvent.Item1, Matches = newEvent.Item2 };
+
+        return StatusCode((int)HttpStatusCode.Created, new ApiResponse { Data = data, Message = "Successfully instantiated event" });
+      }
+      catch (Exception ex)
+      {
+        return ManageException(ex);
+      }
     }
 
     [HttpPost]
@@ -27,6 +72,11 @@ namespace TuPencaUy.Core.API.Controllers
     {
       try
       {
+        if (!string.IsNullOrEmpty(ObtainTenantFromToken()))
+        {
+          return BadRequest(new ApiResponse { Error = true, Message = "You must be logged to a central platform" });
+        }
+
         var createdEvent = _eventService
           .CreateEvent(requestEvent.Name, requestEvent.StartDate, requestEvent.EndDate, requestEvent.Comission, requestEvent.TeamType);
 
@@ -74,6 +124,11 @@ namespace TuPencaUy.Core.API.Controllers
     {
       try
       {
+        if (!string.IsNullOrEmpty(ObtainTenantFromToken()))
+        {
+          return BadRequest(new ApiResponse { Error = true, Message = "You must be logged to a central platform" });
+        }
+
         var createdSport = _eventService.CreateSport(sport.Name, sport.Tie, sport.ExactPoints, sport.PartialPoints);
 
         return StatusCode((int)HttpStatusCode.Created, new ApiResponse { Data = createdSport, Message = "Successfully created sport" });
@@ -111,6 +166,11 @@ namespace TuPencaUy.Core.API.Controllers
     [HttpPost("Team")]
     public IActionResult CreateTeam([FromBody] CreateTeamRequest team)
     {
+      if (!string.IsNullOrEmpty(ObtainTenantFromToken()))
+      {
+        return BadRequest(new ApiResponse { Error = true, Message = "You must be logged to a central platform" });
+      }
+
       try
       {
         var createdTeam = _eventService.CreateTeam(team.Name, team.Logo, team.Sport, team.TeamType);
@@ -152,6 +212,11 @@ namespace TuPencaUy.Core.API.Controllers
     {
       try
       {
+        if (!string.IsNullOrEmpty(ObtainTenantFromToken()))
+        {
+          return BadRequest(new ApiResponse { Error = true, Message = "You must be logged to a central platform" });
+        }
+
         var createdMatch = _eventService
           .CreateMatch(match.EventId, match.FirstTeam, match.SecondTeam, match.FirstTeamScore, match.SecondTeamScore, match.Sport, match.Date);
 
@@ -285,6 +350,11 @@ namespace TuPencaUy.Core.API.Controllers
     {
       try
       {
+        if (!string.IsNullOrEmpty(ObtainTenantFromToken()))
+        {
+          return BadRequest(new ApiResponse { Error = true, Message = "You must be logged to a central platform" });
+        }
+
         _eventService.DeleteMatch(idMatch);
 
         var successResponse = new ApiResponse
@@ -305,6 +375,11 @@ namespace TuPencaUy.Core.API.Controllers
     {
       try
       {
+        if (!string.IsNullOrEmpty(ObtainTenantFromToken()))
+        {
+          return BadRequest(new ApiResponse { Error = true, Message = "You must be logged to a central platform" });
+        }
+
         _eventService.DeleteTeam(idTeam);
 
         var successResponse = new ApiResponse
@@ -325,6 +400,11 @@ namespace TuPencaUy.Core.API.Controllers
     {
       try
       {
+        if (!string.IsNullOrEmpty(ObtainTenantFromToken()))
+        {
+          return BadRequest(new ApiResponse { Error = true, Message = "You must be logged to a central platform" });
+        }
+
         _eventService.DeleteSport(idSport);
 
         var successResponse = new ApiResponse
@@ -345,6 +425,11 @@ namespace TuPencaUy.Core.API.Controllers
     {
       try
       {
+        if (!string.IsNullOrEmpty(ObtainTenantFromToken()))
+        {
+          return BadRequest(new ApiResponse { Error = true, Message = "You must be logged to a central platform" });
+        }
+
         _eventService.DeleteEvent(idEvent);
 
         var successResponse = new ApiResponse
@@ -365,6 +450,11 @@ namespace TuPencaUy.Core.API.Controllers
     {
       try
       {
+        if (!string.IsNullOrEmpty(ObtainTenantFromToken()))
+        {
+          return BadRequest(new ApiResponse { Error = true, Message = "You must be logged to a central platform" });
+        }
+
         var m = _eventService.ModifyMatch(idMatch, match.FirstTeam, match.SecondTeam, match.Date, match.FirstTeamScore, match.SecondTeamScore, match.Sport);
 
         var successResponse = new ApiResponse
@@ -386,6 +476,11 @@ namespace TuPencaUy.Core.API.Controllers
     {
       try
       {
+        if (!string.IsNullOrEmpty(ObtainTenantFromToken()))
+        {
+          return BadRequest(new ApiResponse { Error = true, Message = "You must be logged to a central platform" });
+        }
+
         var t = _eventService.ModifyTeam(idTeam, team.Name, team.Logo, team.TeamType, team.Sport);
 
         var successResponse = new ApiResponse
@@ -407,6 +502,11 @@ namespace TuPencaUy.Core.API.Controllers
     {
       try
       {
+        if (!string.IsNullOrEmpty(ObtainTenantFromToken()))
+        {
+          return BadRequest(new ApiResponse { Error = true, Message = "You must be logged to a central platform" });
+        }
+
         var s = _eventService.ModifySport(idSport, sport.Name, sport.Tie, sport.ExactPoints, sport.PartialPoints);
 
         var successResponse = new ApiResponse
@@ -428,6 +528,11 @@ namespace TuPencaUy.Core.API.Controllers
     {
       try
       {
+        if (!string.IsNullOrEmpty(ObtainTenantFromToken()))
+        {
+          return BadRequest(new ApiResponse { Error = true, Message = "You must be logged to a central platform" });
+        }
+
         var e = _eventService.ModifyEvent(idEvent, ev.Name, ev.StartDate, ev.EndDate, ev.Comission, ev.TeamType);
 
         var successResponse = new ApiResponse
